1786: chore: increase ps_retries and silence log once logged r=Abhinandan-Purkait a=Abhinandan-Purkait

- Increase the ps_retries to 300.
- Silently retry.

+++ io-engine/src/persistent_store.rs
@@ -36,7 +36,7 @@ pub struct PersistentStoreBuilder {
     /// Operation timeout.
     timeout: Duration,
     /// Number of operation retries.
-    retries: u8,
+    retries: u16,
 }
 
 impl Default for PersistentStoreBuilder {
@@ -79,7 +79,7 @@ impl PersistentStoreBuilder {
     }
 
     /// Sets number of operation retries.
-    pub fn with_retries(mut self, retries: u8) -> Self {
+    pub fn with_retries(mut self, retries: u16) -> Self {
         self.retries = retries;
         self
     }
@@ -101,7 +101,7 @@ pub struct PersistentStore {
     /// Operation timeout.
     timeout: Duration,
     /// Number of operation retries.
-    retries: u8,
+    retries: u16,
 }
 
 /// Persistent store global instance.
@@ -304,7 +304,7 @@ impl PersistentStore {
     }
 
     /// Gets the number of operation retries.
-    pub fn retries() -> u8 {
+    pub fn retries() -> u16 {
         Self::instance().lock().retries
     }
